pyproject.toml
[project]
name = "first-project"
version = "0.1.0"
description = ""
authors = [
    {name = "dima",email = "prostodima992@gmail.com"}
]
packages = [{include = "first_project"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
[tool.ruff]
line-length = 100
